Intelli-Vision/requirements.txt
streamlit
ultralytics
pygobject
gtts
playsound
pyttsx3
pynmea2
picamera2
pynmea2
google-generativeai
google-cloud-aiplatform
tflite-runtime
python-dotenv
geocoder
python-telegram-bot